Ref: Rename outgoing -> mailer
Ref: Move mailing -> mailer
